narad-agent-backend/requirements.txt
langchain==0.1.0
langchain_community==0.0.13
langgraph==0.0.20
langchain-openai==0.0.5
langchain_groq
pydantic>=2.0,<3.0
python-dotenv==1.0.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Future dependencies (for upcoming phases)
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9